+++ packages/cli/src/commands/sandbox/sandbox-secret/sandbox_secret_get_command.test.ts
@@ -34,10 +34,10 @@ void describe('sandbox secret get command', () => {
   );
 
   const sandboxIdResolver: SandboxBackendIdResolver = {
-    resolve: () =>
+    resolve: (nameOverride?: string) =>
       Promise.resolve({
         namespace: testBackendId,
-        name: testSandboxName,
+        name: nameOverride || testSandboxName,
         type: 'sandbox',
       }),
   } as SandboxBackendIdResolver;
@@ -66,6 +66,26 @@ void describe('sandbox secret get command', () => {
       {
         namespace: testBackendId,
         name: testSandboxName,
+        type: 'sandbox',
+      },
+      testSecretIdentifier,
+    ]);
+
+    assert.equal(printMock.mock.callCount(), 1);
+    assert.deepStrictEqual(
+      printMock.mock.calls[0].arguments[0],
+      format.record(testSecret)
+    );
+  });
+
+  void it('gets secret for named sandbox', async () => {
+    await commandRunner.runCommand(`get ${testSecretName} --name anotherName`);
+
+    assert.equal(secretGetMock.mock.callCount(), 1);
+    assert.deepStrictEqual(secretGetMock.mock.calls[0].arguments, [
+      {
+        namespace: testBackendId,
+        name: 'anotherName',
         type: 'sandbox',
       },
       testSecretIdentifier,