+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aws-amplify/backend-auth
 
+## 0.5.0-beta.7
+
+### Patch Changes
+
+- @aws-amplify/backend-output-storage@0.4.0-beta.4
+- @aws-amplify/auth-construct-alpha@0.6.0-beta.7
+
 ## 0.5.0-beta.6
 
 ### Patch Changes